requirements.txt
streamlit
plotly
yfinance
requests
streamlit-autorefresh
groq
python-dotenv
